+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 * Added `Kreait\Firebase\Auth\User::getEmail()`
+* Added `Kreait\Firebase\Auth\User::hasVerifiedEmail()`
 * Added `Kreait\Firebase\Auth::sendPasswordResetEmail($userOrEmail)`
   ([Documentation](https://firebase-php.readthedocs.io/en/latest/user-management.html#send-a-password-reset-email))